Move STPStringUtils to StripePaymentSheet
Moves string utility functions from StripePaymentsUI to StripePaymentSheet.
