+++ StripePaymentSheet/StripePaymentSheet/Source/Categories/STPPaymentIntentConfirmParams+PaymentSheet.swift
@@ -6,7 +6,7 @@
 //
 
 import Foundation
-@_spi(STP) import StripePaymentsUI
+@_spi(STP) import StripePayments
 
 extension STPPaymentIntentConfirmParams {
     var nonnil_paymentMethodOptions: STPConfirmPaymentMethodOptions {
