+++ StripePaymentSheet/StripePaymentSheet/Source/Categories/STPPaymentMethodParams+PaymentSheet.swift
@@ -7,7 +7,6 @@
 //
 
 import Foundation
-@_spi(STP) import StripePaymentsUI
 
 extension STPPaymentMethodParams {
     var nonnil_billingDetails: STPPaymentMethodBillingDetails {
