+++ StripePaymentSheet/StripePaymentSheet/Source/Helpers/STPStringUtils.swift
@@ -1,18 +1,15 @@
 //
-//  STPStringUtils.swift
+//  PaymentSheetStringUtils.swift
 //  StripePaymentSheet
-//
-//  Created by Brian Dorfman on 9/7/16.
-//  Copyright © 2016 Stripe, Inc. All rights reserved.
 //
 
 import Foundation
 @_spi(STP) import StripeCore
-@_spi(STP) import StripePaymentsUI
 
 typealias STPTaggedSubstringCompletionBlock = (String?, NSRange) -> Void
 typealias STPTaggedSubstringsCompletionBlock = (String, [String: NSValue]) -> Void
-extension STPStringUtils {
+
+class PaymentSheetStringUtils: NSObject {
     enum Error: Swift.Error {
         case tagMissing
         case tagsNotApplied
@@ -29,7 +26,7 @@ extension STPStringUtils {
     ///   - tag:        The tag to search for.
     ///   - completion: The string with the named tag removed and the range of the
     /// substring it covered.
-    @objc(parseRangeFromString:withTag:completion:) class func parseRange(
+    class func parseRange(
         from string: String,
         withTag tag: String,
         completion: STPTaggedSubstringCompletionBlock
@@ -76,7 +73,7 @@ extension STPStringUtils {
     /// substrings they covered (wrapped in NSValue)
     /// @warning Doesn't currently support overlapping tag ranges because that's
     /// complicated and we don't need it at the moment.
-    @objc(parseRangesFromString:withTags:completion:) class func parseRanges(
+    class func parseRanges(
         from string: String,
         withTags tags: Set<String>,
         completion: STPTaggedSubstringsCompletionBlock
@@ -172,7 +169,7 @@ extension STPStringUtils {
     class func applyLinksToString(template: String, links: [String: URL]) -> NSMutableAttributedString {
         let formattedString = NSMutableAttributedString()
         var numberOfLinksApplied = 0
-        STPStringUtils.parseRanges(from: template, withTags: Set<String>(links.keys)) { string, matches in
+        PaymentSheetStringUtils.parseRanges(from: template, withTags: Set<String>(links.keys)) { string, matches in
             formattedString.append(NSAttributedString(string: string))
             for (tag, range) in matches {
                 guard range.rangeValue.location != NSNotFound else {
@@ -202,5 +199,87 @@ extension STPStringUtils {
             stpAssertionFailure("Failed to apply links '\(links)' to '\(template)'")
         }
         return formattedString
+    }
+
+    // This code was adapted from Stripe.js
+    /// Reformats an expiration date with a four-digit year to one with a two digit year.
+    /// Ex: `01/2021` to `01/21`.
+    static let expirationDateStringRegex: NSRegularExpression = {
+        return try! NSRegularExpression(
+            pattern: "^(\\d{2}\\D{1,3})(\\d{1,4})?",
+            options: []
+        )
+    }()
+
+    class func expirationDateString(
+        from string: String?
+    )
+        -> String?
+    {
+        guard let string = string else {
+            return nil
+        }
+        guard
+            let result = expirationDateStringRegex.matches(
+                in: string,
+                options: [],
+                range: NSRange(location: 0, length: string.count)
+            ).first
+        else {
+            return string
+        }
+        if result.numberOfRanges > 1 && result.range(at: 2).length == 4 {
+            // If a 4-digit year was pasted, shorten it to the last 2 digits
+            var range = result.range(at: 2)
+            range.length = 2
+            range.location = (range.location) + 2
+            let month = (string as NSString).substring(with: result.range(at: 1))
+            let year = (string as NSString).substring(with: range)
+            return "\(month)\(year)"
+        }
+
+        return string
+    }
+
+    static let stringMayContainExpirationDateRegex: NSRegularExpression? = {
+        return try! NSRegularExpression(
+            pattern: "^(\\d{2}\\D{1,3})(\\d{1,4})?",
+            options: []
+        )
+    }()
+
+    /// Returns YES if the string is likely to contain something formatted similar to an expiration date.
+    /// It doesn't confirm that the expiration date is valid, or that it is even a date.
+    class func stringMayContainExpirationDate(_ string: String?) -> Bool {
+        let result = stringMayContainExpirationDateRegex?.matches(
+            in: string ?? "",
+            options: [],
+            range: NSRange(location: 0, length: string?.count ?? 0)
+        ).first
+        return result != nil && (result?.numberOfRanges ?? 0) > 0
+    }
+
+    static let slashFormattedExpirationDateRegex: NSRegularExpression? = {
+        return try? NSRegularExpression(pattern: #"\b(\d{2})/(\d{2,4})\b"#)
+    }()
+
+    /// Returns a sanitized expiration date (in "0101" format) from a string.
+    /// This differs from the existing expiration date parser, as it only looks for dates formatted by slashes.
+    /// This is only intended for OCR use, as we'll often get messy strings like "Exp 01/20 Verification Code 123"
+    class func sanitizedExpirationDateFromOCRString(_ string: String) -> String? {
+        guard let regex = slashFormattedExpirationDateRegex,
+              let match = regex.firstMatch(in: string, range: NSRange(string.startIndex..., in: string)),
+              match.numberOfRanges >= 3,
+              let monthRange = Range(match.range(at: 1), in: string),
+              let yearRange = Range(match.range(at: 2), in: string) else {
+            return nil
+        }
+
+        let monthStr = String(string[monthRange])
+        let yearStr = String(string[yearRange])
+
+        // Combine month and year into MMYY format
+        let sanitizedExpiration = monthStr + yearStr.suffix(2)
+        return sanitizedExpiration
     }
 }
